Move stablecoins pull to directory
