+++ scripts/python_deps.sh
@@ -24,6 +24,13 @@ requires-python = ">=3.12"
 EOF
 
 
-# Exctract all lines between the /START/,/END/ patterns (inclusive).
-# These lines contain the dependencies list.
-awk '/^dependencies =/,/^]/' pyproject.toml  >> dummydeps/pyproject.toml
+# Extract all lines between the /START/,/END/ patterns (inclusive), where
+#
+#   START = '^dependencies'
+#     END = '^]'
+#
+# These lines contain the dependencies in pyproject.toml.
+awk '/^dependencies =/,/^]/' pyproject.toml  | grep -v "sugar" >> dummydeps/pyproject.toml
+
+# NOTE: In the above command we remove the "sugar" dependency. The sugar-sdk is not
+# published to PyPI so it is installed manually in the Dockerfile.
